adicionadas funções no contexto de campanhas
