+++ src/App.js
@@ -2,17 +2,20 @@ import './App.css';
 import Routers from './routers';
 import Header from './components/header';
 import Footer from './components/footer';
-import MenuProvider from './context/context'
+import MenuProvider from './context/context';
+import { CampanhaProvider } from './context/CampanhaContext';
 import { BrowserRouter } from 'react-router-dom';
 
 function App() {
   return (
     <MenuProvider>
+      <CampanhaProvider>
       <BrowserRouter>
         <Header />
         <Routers />
         <Footer />
       </BrowserRouter>
+      </CampanhaProvider>
     </MenuProvider>
 
   );
